Add left and right pseudo inverse using INDArrays
Add methods to compute the left and right pseudo inverse by simple
matrix multiplications directly in ND4J according to
https://en.wikipedia.org/wiki/Moore%E2%80%93Penrose_inverse#Definition

The method makes use of InvertMatrix.invert(..) which at the moment
uses commons math so this is not a 100% ND4J based implementation but
as soon as someone implements an invert method in ND4J using blas this
should be fine.

Tests for left and right inverse are added in TestInvertMatrices.java
for working and non working (singular matrix) cases.

See also: #2858